+++ src/xml/xml_native_writer.cc
@@ -1056,6 +1056,7 @@ void mjXWriter::Option(XMLElement* root) {
     WRITEDSBL("eulerdamp",      mjDSBL_EULERDAMP)
     WRITEDSBL("autoreset",      mjDSBL_AUTORESET)
     WRITEDSBL("nativeccd",      mjDSBL_NATIVECCD)
+    WRITEDSBL("island",         mjDSBL_ISLAND)
 #undef WRITEDSBL
 
 #define WRITEENBL(NAME, MASK) \
@@ -1066,7 +1067,6 @@ void mjXWriter::Option(XMLElement* root) {
     WRITEENBL("fwdinv",         mjENBL_FWDINV)
     WRITEENBL("invdiscrete",    mjENBL_INVDISCRETE)
     WRITEENBL("multiccd",       mjENBL_MULTICCD)
-    WRITEENBL("island",         mjENBL_ISLAND)
 #undef WRITEENBL
   }