+++ test/experimental/usd/mjcPhysics/mjc_physics_scene_test.cc
@@ -180,7 +180,6 @@ TEST_F(MjcPhysicsSceneTest, TestDefaults) {
                                                    mjENBL_INVDISCRETE);
   EXPECT_ENABLE_FLAG_USD_FALLBACK_EQ_MODEL_DEFAULT(MultiCCDFlag,
                                                    mjENBL_MULTICCD);
-  EXPECT_ENABLE_FLAG_USD_FALLBACK_EQ_MODEL_DEFAULT(IslandFlag, mjENBL_ISLAND);
 
   mj_deleteModel(default_model);
   mj_deleteSpec(empty_spec);